Refactoring of computing parts of class EccwExplore
Computing is genralised, with usage of 'numpy' elements.
The consuming part (apply _function_to_root to a 3D grid)
is now using a map function, which may allow usage of
parallel dispatch.
Core modules 'multiprocessing' and 'concurrent' fail at
pickling the class itself. A solution seems to be the
third party module 'pathos', which proposes its own
pickling tool (dill), with a fork of multiprocessing.